third_party/blink/web_tests/platform/mac-mac10.10/http/tests/devtools/console/console-format-style-expected.txt

CONSOLE MESSAGE: line 14: %cBlue!. color: blue;
CONSOLE MESSAGE: line 15: %cBlue! %cRed! color: blue; color: red;
CONSOLE MESSAGE: line 16: %cBlue!
%cRed! color: blue; color: red;
CONSOLE MESSAGE: line 17: %cwww.google.com color: blue
CONSOLE WARNING: line 18: %cwww.google.com color: blue; background: blue
CONSOLE ERROR: line 19: %cwww.google.com color: blue; background: blue
CONSOLE MESSAGE: line 20: [30m1[31m2[90m3[91m4[40m5[41m6[100m7[101m8
CONSOLE MESSAGE: line 21: [30m%d[31m%f[90m%s[91m%d[40m%f[41m%s[100m%d[101m%f 1 1.1 a 2 2.2 b 3 3.3
Tests that console logging dumps properly styled messages.

console-format-style.js:14 Blue!.
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: blue;
console-format-style.js:15 Blue! Red!
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: blue;
Styled text #1: contain: paint; display: inline-block; max-width: 100%; color: red;
console-format-style.js:16 Blue!Red!
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: blue;
Styled text #1: NO STYLES DEFINED
Styled text #2: contain: paint; display: inline-block; max-width: 100%; color: blue;
Styled text #3: contain: paint; display: inline-block; max-width: 100%; color: red;
console-format-style.js:17 www.google.com
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: blue;
Styled text #1: -webkit-text-stroke: 0px !important; text-decoration: underline !important; color: rgb(84, 84, 84) !important; background-color: rgb(255, 255, 255) !important;
console-format-style.js:18 www.google.com
onload @ console-format-style.js:18
(anonymous) @ console-format-style.js:22
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: blue; background: blue;
Styled text #1: -webkit-text-stroke: 0px !important; text-decoration: underline !important; color: rgb(84, 84, 84) !important; background-color: rgb(255, 251, 229) !important;
console-format-style.js:19 www.google.com
onload @ console-format-style.js:19
(anonymous) @ console-format-style.js:22
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: blue; background: blue;
Styled text #1: -webkit-text-stroke: 0px !important; text-decoration: underline !important; color: rgb(84, 84, 84) !important; background-color: rgb(255, 240, 240) !important;
console-format-style.js:20 12345678
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: black;
Styled text #1: contain: paint; display: inline-block; max-width: 100%; color: red;
Styled text #2: contain: paint; display: inline-block; max-width: 100%; color: darkgray;
Styled text #3: contain: paint; display: inline-block; max-width: 100%;
Styled text #4: contain: paint; display: inline-block; max-width: 100%; background: black;
Styled text #5: contain: paint; display: inline-block; max-width: 100%; background: red;
Styled text #6: contain: paint; display: inline-block; max-width: 100%; background: darkgray;
Styled text #7: contain: paint; display: inline-block; max-width: 100%;
console-format-style.js:21 11.1a22.2b33.3
Styled text #0: contain: paint; display: inline-block; max-width: 100%; color: black;
Styled text #1: contain: paint; display: inline-block; max-width: 100%; color: red;
Styled text #2: contain: paint; display: inline-block; max-width: 100%; color: darkgray;
Styled text #3: contain: paint; display: inline-block; max-width: 100%;
Styled text #4: contain: paint; display: inline-block; max-width: 100%; background: black;
Styled text #5: contain: paint; display: inline-block; max-width: 100%; background: red;
Styled text #6: contain: paint; display: inline-block; max-width: 100%; background: darkgray;
Styled text #7: contain: paint; display: inline-block; max-width: 100%;
